# openapi-cleaned.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLARITY Digital Twin Platform",
    "description": "Production AWS-native health data platform with comprehensive API endpoints",
    "version": "0.2.0",
    "x-logo": {
      "url": "https://clarity.novamindnyc.com/logo.png",
      "altText": "CLARITY Logo"
    },
    "contact": {
      "name": "CLARITY Support",
      "email": "support@clarity.novamindnyc.com",
      "url": "https://clarity.novamindnyc.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://clarity.novamindnyc.com/license"
    }
  },
  "paths": {
...
    "/api/v1/insights/{insight_id}": {
      "get": {
        "tags": [
          "ai-insights",
          "ai-insights"
        ],
        "summary": "Get Cached Insight",
        "description": "Retrieve a previously generated insight by ID",
        "operationId": "get_insight_api_v1_insights__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/insights/history/{user_id}": {
      "get": {
        "tags": [
          "ai-insights",
          "ai-insights"
        ],
        "summary": "Get Insight History",
        "description": "Retrieve insight generation history for a user",
        "operationId": "get_insight_history_api_v1_insights_history__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/insights/status": {
      "get": {
        "tags": [
          "ai-insights",
          "ai-insights"
        ],
        "summary": "Service Health Status",
        "description": "Check the health status of the Gemini insights service",
        "operationId": "get_service_status_api_v1_insights_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HealthInsightResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "narrative": {
            "type": "string",
            "title": "Narrative",
            "description": "Human-readable health narrative"
          },
          "key_insights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Insights",
            "description": "Key insights extracted"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Actionable recommendations"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "description": "Confidence in the insights (0-1)"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At",
            "description": "Timestamp of generation"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "narrative",
          "key_insights",
          "recommendations",
          "confidence_score",
          "generated_at"
        ],
        "title": "HealthInsightResponse",
        "description": "Response containing generated health insights."
      },
...
      "InsightGenerationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/HealthInsightResponse"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "title": "InsightGenerationResponse",
        "description": "Response for insight generation."
      },
      "InsightHistoryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "type": "object",
            "title": "Data"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "title": "InsightHistoryResponse",
        "description": "Response for insight history."
      },
...
      "ServiceStatusResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "type": "object",
            "title": "Data"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "title": "ServiceStatusResponse",
        "description": "Response for service status."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": true
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://clarity-alb-1762715656.us-east-1.elb.amazonaws.com",
      "description": "Production server (AWS ALB)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
...
}